リザルト画面でもプレイヤータグを使用するため、Co Areaのプレイヤータグをコンポーネント化した
