+++ vue/src/components/CoArea.vue
@@ -1,88 +1,34 @@
 <template>
   <section class="co-area">
-    <!-- 自分 -->
     <div class="player-wrapper">
-      <article
-        class="player"
-        :class="{ selected: isSelected(player.playerName) }"
-      >
-        <figure class="player-icon">
-          <img :src="RoleList[player.playerRole.roleName]" alt="" />
-        </figure>
-        <figure class="co-icon">
-          <img
-            class="co-icon__img"
-            :src="this.coRoleList[this.coRole]"
-            :alt="this.coRole"
-          />
-        </figure>
-        <span class="player-name me">{{ player.playerName }}</span>
-      </article>
-
-      <!-- 自分以外のプレイヤー -->
-      <article
-        class="player"
-        :class="{ selected: isSelected(val.name) }"
+      <!-- 自分 -->
+      <Player
+        :playerName="player.playerName"
+        :roleName="player.playerRole.roleName"
+        :coRole="this.coRole"
+        :selectedPlayers="selectedPlayers"
+        :isMe="true"
+      />
+      <!-- 他のプレイヤー -->
+      <Player
+        :playerName="val.name"
+        :roleName="val.role.roleName"
+        :coRole="'村人'"
+        :selectedPlayers="selectedPlayers"
         v-for="(val, key) in otherPlayerList"
         :key="key"
-      >
-        <figure class="player-icon">
-          <img :src="RoleList[val.role.roleName]" :alt="val.role.roleName" />
-        </figure>
-        <figure class="co-icon">
-          <img class="co-icon__img" :src="coRoleList['村人']" alt="" />
-        </figure>
-        <span class="player-name">{{ val.name }}</span>
-      </article>
+      />
     </div>
   </section>
 </template>
 
 <script>
+import Player from "@/components/Player.vue";
+
 export default {
   name: "CoArea",
   props: ["otherPlayerList", "player", "coRole", "selectedPlayers"],
-  data() {
-    return {
-      RoleList: {
-        不明: require("../assets/images/chara-icon/unknown.png"),
-        人狼: require("../assets/images/chara-icon/jinroh.png"),
-        村人: require("../assets/images/chara-icon/murabito.png"),
-        占い師: require("../assets/images/chara-icon/uranaishi.png"),
-        怪盗: require("../assets/images/chara-icon/kaito.png"),
-        狂人: require("../assets/images/chara-icon/kyojin.png"),
-        吊り人: require("../assets/images/chara-icon/tsuribito.png"),
-      },
-      coRoleList: {
-        不明: require("../assets/images/fukidashi/fumei.png"),
-        人狼: require("../assets/images/fukidashi/jinroh.png"),
-        村人: require("../assets/images/fukidashi/murabito.png"),
-        占い師: require("../assets/images/fukidashi/uranaishi.png"),
-        怪盗: require("../assets/images/fukidashi/kaito.png"),
-        狂人: require("../assets/images/fukidashi/kyojin.png"),
-        吊り人: require("../assets/images/fukidashi/tsuribito.png"),
-      },
-    };
-  },
-  methods: {
-    isSelected: function (name) {
-      // 投票ページ以外ではselectedPlayersは存在しないのでfalseを返す
-      if (!this.selectedPlayers) {
-        return false;
-      }
-
-      let selectedPlayersName = [];
-      this.selectedPlayers.forEach((player) => {
-        selectedPlayersName.push(player.name);
-      });
-
-      if (selectedPlayersName.includes(name)) {
-        return true;
-      } else {
-        return false;
-      }
-    },
-  },
+  components: { Player },
 };
 </script>
 
@@ -95,86 +41,6 @@ export default {
     display: flex;
     justify-content: space-between;
     padding: 0 1rem;
-
-    .player {
-      display: grid;
-      grid-template-columns: minmax(80px, 100px) 1fr;
-      column-gap: 16px;
-      align-items: center;
-      width: 30%;
-
-      .player-icon {
-        grid-row: 1/3;
-        grid-column: 1/2;
-        width: 100px;
-        height: 100px;
-
-        img {
-          width: 100%;
-          height: auto;
-          border-radius: 10px;
-        }
-      }
-
-      .co-icon {
-        grid-row: 1/2;
-        grid-column: 2/3;
-        width: 50px;
-        height: 50px;
-
-        .co-icon__img {
-          width: 100%;
-          height: auto;
-        }
-      }
-
-      .player-name {
-        grid-row: 2/3;
-        grid-column: 2/3;
-      }
-
-      .me {
-        text-decoration: underline;
-      }
-    }
-
-    .player.selected {
-      .player-icon {
-        position: relative;
-
-        img {
-          filter: grayscale(100%);
-        }
-
-        &::after {
-          position: absolute;
-          left: 0;
-          width: calc(100% * 1.414);
-          height: 2px;
-          content: "";
-          background-color: red;
-          transform: rotateZ(45deg);
-          transform-origin: left top;
-        }
-
-        &::before {
-          position: absolute;
-          bottom: 0;
-          left: 0;
-          z-index: 1;
-          width: calc(100% * 1.414);
-          height: 2px;
-          content: "";
-          background-color: red;
-          transform: rotateZ(-45deg);
-          transform-origin: left top;
-        }
-      }
-
-      .co-icon__img {
-        filter: grayscale(100%);
-      }
-    }
   }
 }
 
@@ -188,21 +54,6 @@ export default {
       flex-direction: column;
       row-gap: 1rem;
       align-items: center;
-
-      .player {
-        width: 100%;
-        max-width: 320px;
-
-        .player-icon {
-          width: 80px;
-          height: 80px;
-        }
-
-        .co-icon {
-          width: 50px;
-          height: 50px;
-        }
-      }
     }
   }
 }
